Fix direct access to token variable
