+++ backend/src/routes/liveness.ts
@@ -1,5 +1,4 @@
 /* Copyright Contributors to the Open Cluster Management project */
-import { readFileSync } from 'fs'
 import { constants, Http2ServerRequest, Http2ServerResponse } from 'http2'
 import { Agent } from 'https'
 import { FetchError } from 'node-fetch'
@@ -7,6 +6,7 @@ import { fetchRetry } from '../lib/fetch-retry'
 import { logger } from '../lib/logger'
 import { respondInternalServerError, respondOK } from '../lib/respond'
 import { getOauthInfoPromise } from './oauth'
+import { getServiceAccountToken } from './serviceAccountToken'
 const { HTTP2_HEADER_AUTHORIZATION } = constants
 
 // The kubelet uses liveness probes to know when to restart a container.
@@ -26,33 +26,13 @@ export function setDead(): void {
     }
 }
 
-export function getServiceAccountToken(): string {
-    if (serviceAccountToken === undefined) {
-        try {
-            serviceAccountToken = readFileSync('/var/run/secrets/kubernetes.io/serviceaccount/token', 'utf-8')
-        } catch (err: unknown) {
-            serviceAccountToken = process.env.TOKEN
-            if (!serviceAccountToken) {
-                if (err instanceof Error) {
-                    logger.error('Error reading service account token', err && err.message)
-                } else {
-                    logger.error({ msg: 'Error reading service account token', err: err })
-                }
-                process.exit(1)
-            }
-        }
-    }
-    return serviceAccountToken
-}
-let serviceAccountToken: string
-
 const agent = new Agent({ rejectUnauthorized: false })
 
 export async function apiServerPing(): Promise<void> {
     const msg = 'kube api server ping failed'
     try {
         const response = await fetchRetry(process.env.CLUSTER_API_URL + '/apis', {
-            headers: { [HTTP2_HEADER_AUTHORIZATION]: `Bearer ${serviceAccountToken}` },
+            headers: { [HTTP2_HEADER_AUTHORIZATION]: `Bearer ${getServiceAccountToken()}` },
             agent,
         })
         if (response.status !== 200) {
